support fractional epochs in trainer
Summary:
When sampling users with replacement from a very large user pool, it is useful to train for fewer rounds than `(num_total_users//num_users_per_round)`.
It is especially helpful to reduce the noise needed to achieve a certain epsilon value since the sampling rate decreases but the number of steps doesn't.

This commit adds the above support via config by allowing the user to specify a fractional epoch.

Reviewed By: JohnlNguyen

Differential Revision: D32117471

fbshipit-source-id: f76824ebb88ee446cc7e852e030e6ffdbb247112