Move optimizer_test_utils to optimizers directory (#18)
Summary:
Pull Request resolved: https://github.com/facebookresearch/FLSim/pull/18

it is currently located at the top-level `tests` directory. However the top-level `tests` directory does not really make  sense as each component is organized into its dedicated directory. `optimizer_test_utils.py` belongs to the `optimizer` directory in that sense. In this diff, we move the file to the `optimizer` directory and fixes the reference.

Differential Revision: D32241821

fbshipit-source-id: 77051d22b44a5a136518183ca279234f06e961ac