Stats Collector refactor
Summary:
Channel stats collector refactor
The following are the main changes in this refactor:
1. We do not have trainer own the stats collector anymore and attach a channel to it. The channel owns it from now on, which enables a dedicated collector per channel and doesn't burden the trainer. Hydra configs for the measurement reporting are now within the channel configs.
2.  Also modified the channel APIs a bit to ensure that we do not worry too much about the stats collection when implementing a new channel and do not repeat code too much:
a. `_calc_message_size()` is added as an abstract method and has to be implemented in each channel.
b. With this change, it is now sufficient to just inherit from `IdentityChannel` which just has `_during_transmission_client_to_server()` calling message calculation. This avoids redefinition of `_during_transmission_client_to_server()` in every channel if it just calculates message size and doesn't do anything else.
3. This is a refactor, so overall operation does not change. We have some code length savings.

There might be other parts to this refactor, will change later as need be.

Reviewed By: JohnlNguyen

Differential Revision: D30902962

fbshipit-source-id: d86faf1b3dbc98fd35ab522f6672bdf266ee3d4e