Sync Server Intergration
Summary:
## Summary
This diff replaces the aggregator/reducer with the server. It integrates the server concept with the sync trainer. For more context about the server, see https://fb.quip.com/AbwcAzKOCZwW.

## What?
1. Replace aggregator/reducer/active user selector with server
2. Remove the need for PrivateSyncTrainer. With the server, a user can do the following for DP training:
```
    "_base_": "base_sync_trainer",
    "server": {
      "_base_": "base_sync_dp_server",
      "privacy_setting": {
        "alphas": [2, 3, 4, 6, 7, 8, 9, 10, 20, 30, 40, 60],
        "noise_multiplier": 0.0,
        "clipping_value": 10000.0,
        "target_delta": 1e-6
      },
```
or the following for normal non-dp training
```
      "_base_": "base_sync_trainer",
      "server": {
        "_base_": "base_sync_server",
        "optimizer": {
          "type": "FedAvgWithLR",
          "lr": 1,
          "momentum": 0.
        }
      },
```

3. Fix broken tests because of the change
```
## Note
The server does not have channels. I will do integrate the channel in the next diff in the stack.
```

## How to read this diff
This is a large diff because all the tests I had to fix. The core changes are minimal.

So to read this diff, please pay attention to the following files (in order):
1. sync_trainer.json for new non-DP config
2. sync_trainer_with_dp.json for new DP json config
3. sync_trainer.py
4. trainer_base.py

Reviewed By: karthikprasad

Differential Revision: D32222463

fbshipit-source-id: 0442dfd180ba06b6943784a2f8e8636728e255a7