Move data sharder to hydra
Summary:
TLDR: Sharder currently uses a custom config hack. Fixing it here.

## Details:
Removed the SharderFactory which accepted sharding strategy type and constructed a sharder with applicable args via an if-else ladder. Instead, every sharder is a class of its own, and is constructed directly from config via hydra just like every other flsim component. Since we are using hydra, the schema enforcement and validation is taken care of for free.

## How to review:
1. Review data_sharder.py first. Bulk of the changes are in that file. That said, the changes themselves are not new as such. Much of the code is retained (as shown by the yellow border on LHS), and new code is boiler plate
2. Review test_data_sharder.py to see how this new change will reflect in usage.
3. Review pytext_fl_dataloader.py and the example pytext config json to see how this upstream change will affect pytext.
4. Review config_adaptor.py to ensure I'm adapting correctly.
5. Wait for sandcastle to go green.

Reviewed By: JohnlNguyen

Differential Revision: D32452243

fbshipit-source-id: d69fca16d73905bc1f4fda36b3221117b1b70490