Remove reference to fb.data in core library
Summary:
Currrently, importing `flsim.configs` imports all trainer related configs AND configs for paged dataloader and hive dataset. While the former is fine, the latter is fb specific and should not be exposed in the open source library.

This diff fixes the above by creating `flsim.fb.configs` to for internal use (eg in fblearner, pytext, and baselines). `flsim.configs` can still be used for external use (eg in examples and tutorials)

Reviewed By: JohnlNguyen

Differential Revision: D32482031

fbshipit-source-id: eda2aba0a48b9a23a106f2ac0290499e4d23375b