Fix DP issue after server refactor
Summary:
server refactored was not fully tested, this means that we are missing unit tests.
We used to have tests they were not ported.
https://www.internalfb.com/code/fbsource/fbcode/papaya/toolkit/simulation/reducers/tests/round_reducer_test.py

sync operation fails for any dp use case with multiple workers.

the issue raises from the fact that a float is passed to the sync operation that expects a tensor.

there is a side issue *bad naming*,
aggregator has a properry called weights, this does not represent what it is, i.e. the sum_weights this also fixes this issue as weights is usually a vector of weights as used elsewhere,
i.e.
```
./clients/tests/client_test.py:142:        assertAlmostEqual(weight1, client.weights[0])
```

Reviewed By: JohnlNguyen

Differential Revision: D32582684

fbshipit-source-id: 54176feeba017063a54e1dde7b10b16030a13ed0