Add TopK sparsity and generalize sparse mask channels
Summary:
In this diff, we generalize the sparse mask channels to account for heuristic sparse masks. This allows for 1. random masks as we had previously, as well as 2. top-k sparsity where the lowest abs values are zeroed out.
I did not create a new channel as most of the functionality is very similar between the two, including channel measurement.
If you think we should have top-k mask as separate channel, we can do that.

Reviewed By: pierrestock

Differential Revision: D32194784

fbshipit-source-id: 3f7c5c8df647ff29660840fa3060204e57696156