Internal updates for 1.0
Summary:
Fixing internal opacus dependencies to use the new API

There are 4 components in need for update:

## opacus TARGETS
For opacus itself we need to update buck build rules

## flsim
flsim uses opacus for per-sample DP.
FLSim has its own notion of data processing, so opacus doesn't have access to a `DataLoader` object required to use `PrivacyEngine`. No big deal, we just use `GradSampleModule`, `DPOptimizer` and `RDPAccountant`

## caffe2dp
Caffe2dp uses opacus as a gold standard for tests. Updating it was straighforward

## label dp
Our implementation for Label DP paper was also based on an old Opacus API. Normally we'd just list `opacus==0.14` as the hard requirement. However, in fbcode you have to always depend on main.
So, since the code was experimental and is not used anywhere, I disabled all the tests for it. The code is still there for historical purposes, but essentially dead.
(We could also delete it altogether)

## loop/federated_learning
I have very little insight into what's going on there. Also, this was the only part I wasn't able to check with unit tests.
But it should work - I do the same fix as with flsim (here we also have weird relationship with the data loader)

Reviewed By: Anonymani, ashkan-software

Differential Revision: D32617805

fbshipit-source-id: 12a93c3fbce85a1a4c0c7dbca460cb07b4ec2e46