Add Custom Channel Tutorial
Summary:
First draft of a tutorial to add a custom channel in FLSim.

In this tutorial, you will learn:
- - The structure of a Message used to communicate any relevant information between the server and the clients.
The API of a channel with the example of the IdentityChannel. It implements a pass-through without any modification to the underlying message.
- How to implement a custom channel with the example of the ScalarQuantizationChannel.
- - How to perform a training in FLSim with your custom channel.
How to measure the number of bytes sent from the client to the server and from the server to the client.

Bento NB: N1405454
Colab NB: [here](https://colab.research.google.com/drive/1p6InKd07ourNAGHCBanvkx0fTVdmInbY?usp=sharing)

Reviewed By: sayanghosh, JohnlNguyen

Differential Revision: D32759339

fbshipit-source-id: cc5ec7467474a9eec955470bb42302b0961dee9e