Back out "Improving FLSim Scalability - Chapter 2: Multiple processes per gpu"
Summary:
## What is the problem?
This change forces all runs to use multiple GPUs.

For example,
I'm on a machine with 8 GPUs but only want to use 1 by setting `distributed_world_size=1`. This setting is common for most researchers. They're on a machine with multiple GPUs but would like to run an experiment on 1.

The line
` distributed_world_size // num_gpus * num_gpus` will return 0 as `1 // 16 = 0`.

## Why revert?
1. This blocks FLSim open sourcing as it breaks our example runs on GPU.
2. This code doesn't have unit tests. I will revert for now until we have time to add unit tests for it.

Reviewed By: Anonymani

Differential Revision: D32774598

fbshipit-source-id: 4c8932f711e3e2736ff7ef08b94860a7c5e20616