Improving FLSim Scalability - Chapter 2: Multiple processes per gpu _ 2nd attempt
Summary:
# WhoAmI
I am a member of a series if diffs improving fl-sim scalability. I focus on multiple processes per gpu.

# Why
When models are small, which is usually the case for FL, with the way we are running client updates: i.e.
* parallel between workers
* serial in each worker
having workers = num gpu's results in gpu under utilization
example:
https://pxl.cl/1RGGC

# what
This diff tries to minimize the serialization in each worker
* by shortening the queue
* by running more than one process on each gpu
this way gpu memory can be shared between two workers and e.g. we can have 16 workers on 8 gpu.
e.g.
f304097296
vs
f303886896

# issues
We have to switch the distributed backend to GLOO, it was previously NCCL for which is better performing for cuda: ref: https://mlbench.github.io/2020/09/08/communication-backend-comparison/

https://pxl.cl/1RGHB

The other drawback of GLOO is that it does not support reduce for gpu (only supports reduce_all) that causes small degradation in performance for when DP is deployed.

But, !hey!, with GLOO you get more processes per GPU so we should be able to enjoy it better at the end of the day. The diff tries to do some optimizations for when #workers <= #gpus to use NCCL

# PS
This diff was once submitted and then reverted. The reason was that the diff has some bug that required the run to have equal or more workers than the number of cudas. If for any reason some one wanted one worker on a 8 cuda machine the job would fail.
This new diff fixes the issue and also  D32815377 which is down the stack ads full unit tests to test the new code.

Reviewed By: kmalik22, ashkan-software

Differential Revision: D33033043

fbshipit-source-id: 378dc2d0febb24c0b67db23a4e0c40d5388fa168