add more coverage for multi-proc tests
Summary:
# what
We are not testing multi-proc with different numbers of processes in test_sync_dp_servers.

# so what
A proper multi-proc test should check, that using different workers does not have any result discrepancy. Also corner cases of one process needs to be checked. This diff ads that.

# is that all?
No, so there were some other issues with the current test file. The process start method was "fork" by default, and FB/Meta disables cuda automatically for this. So we were not testing cuda at all, even on a cuda machine.

Changing to the proper "spawn" method, requires some more changes, and that is what all these other changes are about. e.g. when using "spawn" the server could not be duplicated for processes due to torch.generator so we needed a whole refactoring :(

Reviewed By: kmalik22

Differential Revision: D32815377

fbshipit-source-id: 3f43548247ad4c8af30a107f62993c0c3998293b