Update license from Apache-2.0 to BSD 3-clause (#33)
Summary:
Pull Request resolved: https://github.com/facebookresearch/FLSim/pull/33

FLSim was always supposed to be released under BSD 3 clause license. The license headers in the individual files had correctly listed BSD 3 as the license, however, the LICENSE file in the root directory had the text of Apache 2.0.

Subsequent automated commit (c8548ab846fe098431d6477386c512b45252363e) brought consistency to the codebase by changing the headers to match the root LICENSE, and changed everything to Apache 2.0.

This commit fixes the LICENSE issue fully by changing both the headers and the root LICENSE.

Our apologies to the community if this any confusion. Please comment on this PR if you have any questions.

Reviewed By: zertosh

Differential Revision: D33435411

fbshipit-source-id: a1a3f46f4c5ad02cc192b5f622c094ed460f3157