Add data splitting
Summary:
# Background
Given a dataset, we first hold out a **percentage of clients (e.g., 20%) as unparticipating clients (a.k.a eval clients)**, as shown in the right most three columns.
The remaining clients are participating clients (train clients). We refer to this split as inter-client split.

Within each participating client, we hold out a percentage of data (e.g., 20%) as participating validation data, as shown in the upper left block (in orange). The remaining data is the participating training client data, as shown in the lower left block (in blue). We refer to this second split as intra-client split.

{F699508088}

## High Level Algorithm

1. Server broadcasts model to clients
2. Server selects clients to participate in training, the client trains on its train partition (red box)
3. Client sends back to server its update and server aggregates updates
4. After aggregation, server runs validation on train clients validation samples (yellow box)
5. On validation step, unseen eval clients will take a global model then fine-tune on its fine tune data (green box)
6. The eval client then report validation metrics on fine-tune validation set (purple box)
7. We will use the metrics from the purple box data for HP tuning
8. Testing. Train on Test client (black box), compute test metrics on teal box

# What?
* Change `IFLDataProvider` to accommodate for the inter-client split.
* Change `IFLUserData` to accommodate for intra-client split.

# How to review
1. Look at `papaya/toolkit/simulation/flsim/data/data_provider.py` for the interface change
2. Look at `papaya/toolkit/simulation/flsim/trainers/trainer_base.py` and `papaya/toolkit/simulation/flsim/trainers/sync_trainer.py` for the core call site changes
2. All other files are minor call site changes to fix broken tests

Reviewed By: kmalik22

Differential Revision: D33771328

fbshipit-source-id: 7e386fa7441a899ca52ebda8cc47baf2d2c0c06a