compute overflow only for trainable params
Summary:
The current overflow calculation is correct only when all params are trainable. It will not work correctly when some params are frozen.

This diff addresses the above issue.

Note that this diff still continues to assume that the incoming client updates are updates on model's params; this will not account for overflow in elements that are not part of the model parameter set.

Reviewed By: ashkan-software

Differential Revision: D36562244

fbshipit-source-id: ba62b8cd7941b6587f2c08a35ab7d320399e863d