add support for shared qparams with SQ
Summary:
## Context:
The current implementation of Scalar Quantization of model updates allows each client to use its own scale and zero_point. This diff adds support for shared qparams across all clients, which is especially needed for sec agg.

## This diff
* This diff introduces a global_qparams attribute similar to the global_model in the sync servers.
* This global_qparams is used when we want to share qparams across clients, and are computed on the server after every round.
* Introduces SQ sync server to handle quantization related steps on the server without pesky if-else blocks on the main code.
* Quantization is typically the last step on the client before transmission to server and dequantization is the first step on the server. Therefore, scalar quantization channel simulates both of these. However, dequantization is not the first step on the sever with sec agg in the picture. We handle dequantization separately on the server.

## Reference:
This diff is largely influenced by the changes in D33187367 and D36038474

## Misc Note:
In addition to technical correctness, I'd also like to get your comments on the design. Personally, I find creating a new subclass a bit of an overkill, but the alternative of introducing too many unrelated attributed gated by if-else in the main server is way too ugly.

Reviewed By: pierrestock

Differential Revision: D36090799

fbshipit-source-id: 8c8d7d8e5c5e7c26621a59331fe95d8b6efb15d8