Add DP-FTRL to flsim
Summary:
## What?
* Adding DP-FTRL along with FTRL optimizer to FLSim
* The privacy engine is copied from https://github.com/google-research/DP-FTRL. Most notably, `CummuNoiseTorch` and `CummuNoiseEffTorch`
* The `SyncFTRLServer ` wraps logic for tree completion and restart similar to that of in https://github.com/google-research/DP-FTRL/blob/main/main.py#L208

### What is tree completion?
See Appendix D.3.1
TLDR: Run virtual steps in order to have a complete binary tree so we can have lower noise at the expense of a little more privacy loss.
A natural trick to consider is thus to complete the tree with “virtual steps” such that the noise is the smallest.
{F737392429}

### What is tree restart?
See Appendix D.1
https://arxiv.org/pdf/2103.00039.pdf

## Why?
We're using this code to evaluate utility of FTRL compared to DP-FedAvg

## Notebook to compute epsilon
https://fburl.com/anp/6bms12ap

Reviewed By: pierrestock

Differential Revision: D36358239

fbshipit-source-id: e05f984982478c48111ce01ffe4e76b72c88e0aa