Add simulation progress information to Timeline.
Summary:
### Context
`Timeline` specifies the point in time of simulation. There are two ways of specifying the time:
1. Specify (current) `epoch`, (current) `round` from this epoch, and `rounds_per_epoch`.
2. Specify (current) `global_epoch` and `rounds_per_epoch`.

### Why
It is very useful also have information about the simulation progress, i.e. from 0% to 100%, how far are we in the simulation progress? This information can be used for training operations that act differently in the beginning of the training progress vs. the end of the training progress, for example.

### Implementation
Optionally specify `total_epochs` (i.e. it will have default value of 0) when creating `Timeline`. Add method of `progress_fraction` that returns the percentage of the simulation progress we are done in [0, 1].

It is optional to specify `total_epochs` if `progress_fraction()` will never be used, but once we call `progress_fraction()`, we must have already specified `total_epochs`.

Reviewed By: karthikprasad

Differential Revision: D38606636

fbshipit-source-id: b10c8ddcdb731b075c9113d0a95d646f8eb609a5